internship/code/src/main.py
"""
StakingOptimizer main entry point.

This module initializes and runs the StakingOptimizer agent, which helps users
optimize their staking positions through natural language interactions and
scheduled transactions.
"""

import logging
import os
from typing import Dict, Optional

from dotenv import load_dotenv
from langchain.agents import AgentExecutor
from langchain_core.language_models import BaseLLM

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format="%(asctime)s [%(levelname)s] %(message)s",
    handlers=[
        logging.StreamHandler(),
        logging.FileHandler("staking_optimizer.log"),
    ],
)

logger = logging.getLogger(__name__)

def setup_environment() -> Dict[str, str]:
    """Load and validate environment variables.
    
    Returns:
        Dict[str, str]: Dictionary of environment variables
    
    Raises:
        ValueError: If required environment variables are missing
    """
    load_dotenv()
    
    required_vars = ["OPENAI_API_KEY"]
    env_vars = {}
    
    missing_vars = []
    for var in required_vars:
        value = os.getenv(var)
        if not value:
            missing_vars.append(var)
        else:
            env_vars[var] = value
    
    if missing_vars:
        raise ValueError(f"Missing required environment variables: {', '.join(missing_vars)}")
    
    return env_vars

def create_agent(llm: Optional[BaseLLM] = None) -> AgentExecutor:
    """Create and configure the StakingOptimizer agent.
    
    Args:
        llm: Optional language model to use. If None, will create default.
    
    Returns:
        AgentExecutor: Configured agent executor
    """
    raise NotImplementedError("Agent creation not yet implemented")

def main() -> None:
    """Main entry point for the StakingOptimizer agent."""
    try:
        # Load environment variables
        env_vars = setup_environment()
        logger.info("Environment variables loaded successfully")
        
        # Create agent
        agent = create_agent()
        logger.info("Agent created successfully")
        
        logger.info("Agent starting...")
        
    except Exception as e:
        logger.error(f"Failed to start agent: {e}", exc_info=True)
        raise

if __name__ == "__main__":
    main()
