internship/code/pyproject.toml
[project]
name = "staking-optimizer"
version = "0.1.0"
description = "A LangChain-based staking optimizer agent"
requires-python = ">=3.10"
dependencies = [
    "langchain-ritual-toolkit",
    "langchain>=0.3.17",
    "langchain-core>=0.3.34",
    "langchain-openai>=0.3.4",
    "langgraph>=0.2.70",
    "openai>=1.62.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.7.1",
    "web3>=7.8.0",
    "fastapi>=0.115.8",
    "uvicorn>=0.34.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.25.3",
    "httpx>=0.28.1",
]
dev = [
    "black>=25.1.0",
    "isort>=6.0.0",
    "mypy>=1.15.0",
    "ruff>=0.9.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/staking_optimizer"]

[tool.hatch.metadata]
allow-direct-references = true
